Skip Image Export (node) credential validation
The credentials used for Image Export don't have validation logic,
causing over-all provider credential validation to fail. This blocks
provider refresh from the GUI.

In the future, we should add validation to these credentials if the user
provides them (Image Export is an optional feature, so credentials also
need to be optional).