Add VM 'operating_system' attribute as 'operating_system.name'
PowerVC provides an 'operating_system' attribute for VMs. This is
already persisted as an 'advanced_setting', but that makes including it
in reports more difficult. This commit sets this value to
'operating_system.name' in addition to 'advanced_setting' in order to
allow it to be included in VM reports directly as the 'OS : Name'.